+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # OpenStreetMap Tool
 
+**0.4.0:**
+ - Complete rewrite of search result handling to prevent incorrect OSM
+   map links being generated, and bad info given.
+ - New setting: Instruction Oriented Interpretation. Controls the
+   result interpretation instructions given to the LLM. Should help
+   certain models (Hermes Llama 3.1) give more consistent results.
+ - The Instruction Oriented Interpretation setting is also a user
+   valve that can be controlled on a per-chat basis.
+ - Added ability to search for: public transit, schools/universities,
+   libraries, bike rental locations, car rental locations.
+
+**0.3.0:**
+ - Add search for alcohol and cannabis/smart shops.
+ - Increase hotel search radius to 10km.
+
 **0.2.5:**
  - Better accuracy in turn by turn conversations: model encouraged to
    specify the name of the city and country when searching, so it's